Share the AMR mesh drawer between PROsurf and adaptive FC
Move draw_amr_mesh_on_canvas and AxisXform (both from PROAdaptiveFC.cxx,
where they were an improved copy of PROsurf::PlotAMRMesh) into new
inc/PROmeshPlot.h + src/PROmeshPlot.cxx. PROsurf::PlotAMRMesh becomes a
thin wrapper (canvas + shared draw + Print); the adaptive-FC diagnostics
pick the shared symbols up via using-declarations.

The shared drawer is the adaptive-FC version: shallowest-first leaf
ordering and two-TBox fill+outline rendering (works around ROOT dropping
box borders under SetFillColorAlpha), caller-owned canvas so it also
serves multipage PDFs. The per-level fit-count lines from PROsurf's info
box are folded in so no diagnostic content is lost. Only cosmetic effect:
PROsurf's <tag>_amr_mesh.pdf adopts this rendering; adaptive-FC throw
pages gain the fit-count lines. No numeric output is affected.
